requirements.txt
pyyaml==5.4.1
gym==0.26.2
gym[Box2D]==0.26.2
tensorboard_logger==0.1.0
tensorboard==2.6.0
matplotlib==3.4.3
protobuf==3.6.1